작물아 자라라

+++ Assets/01.Scripts/Farm/CropPlant/CropTypeSO.cs
@@ -24,10 +24,10 @@ public class CropTypeSO : ScriptableObject
     //성장 가능한 계절
     public Season growSeason;
     //성장하는데 걸리는 날짜
-    public int growDay;
+    public float growDay;
     //이놈은 날짜를 시간으로 바꾼 값
     [HideInInspector]
-    public int growTime;
+    public float growTime;
 
     //시간에 따른 스프라이트 변화는 다른곳에서 해줘야 함
 }
